requirements.txt
django
djangorestframework
mechanicalsoup
requests
bs4
xlrd
html5lib
pyyaml